refactor: Sortable list moved to its own component file
